\name{alplot-package}
\alias{alplot-package}
\docType{package}
\title{Plot array lists}
\description{Storing data in array list format is more memory-efficient than data frame format. This package enables easy construction and plotting of these array list data structures.}
\details{
\tabular{ll}{Package: \tab alplot\cr
Maintainer: \tab Toby Dylan Hocking <toby.hocking@inria.fr>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Plot array lists\cr
Imports: \tab lattice\cr
Depends: \tab lattice\cr
Suggests: \tab latticeExtra, nlme\cr}
}
\author{Toby Dylan Hocking <toby.hocking@inria.fr>}

\keyword{ package }


