\name{gd}
\alias{gd}
\title{Get data}
\description{Specify variables to query and return}
\usage{
gd(al, vars, terms = list())
}
\arguments{
  \item{al}{The array.}
  \item{vars}{Vector of variable names to return.}
  \item{terms}{Query terms.}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





