\name{plot.arraylist}
\alias{plot.arraylist}
\title{plot arraylist}
\description{Plot an arraylist using efficient lattice panel functions.}
\usage{
plot.arraylist(x, f, lattice.fun = NULL, ...)
}
\arguments{
  \item{x}{The arraylist.}
  \item{f}{lattice plot formula.}
  \item{lattice.fun}{Lattice plot function to call. default will be xyplot.}
  \item{\dots}{Other arguments to the lattice plot function.}
}



\author{Toby Dylan Hocking <toby.hocking@inria.fr>}





